перемещены повторяющие проверки в Before и After
