+++ src/test/java/TestCreateUser.java
@@ -1,25 +1,32 @@
 import api.UserApi;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.RestAssured;
-import io.restassured.response.Response;
 import io.restassured.response.ValidatableResponse;
 import model.CreateUser;
+import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
-import static io.restassured.RestAssured.given;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.IsNull.notNullValue;
 
 public class TestCreateUser {
 
-    String email = "tesov@test.com";
+    String email = "Smelov@lol.ru";
     String password = "12345";
     String name = "Tuta";
+    String accessToken;
     @Before
     public void setUp(){
         RestAssured.baseURI = "https://stellarburgers.nomoreparties.site/";
+    }
+
+    @After
+    public void cleanUp() {
+        if(accessToken != null) {
+            UserApi.deleteUser(accessToken);
+        }
     }
 
     @Test
@@ -27,19 +34,20 @@ public class TestCreateUser {
     public void createUserTest(){
         CreateUser createUser = new CreateUser(email, password,name);
         ValidatableResponse createUserResponse = UserApi.postCreateUser(createUser);
-        String accessToken = createUserResponse.assertThat().statusCode(200).extract().path("accessToken");
+        accessToken = createUserResponse.assertThat().statusCode(200).extract().path("accessToken");
         assertThat(accessToken, notNullValue());
-        ValidatableResponse deleteUserResponse = UserApi.deleteUser(accessToken);
-        deleteUserResponse.assertThat().statusCode(202);
     }
 
     @Test
     @DisplayName("Creating a user that already exists")
     public void createUserSimpleEmailTest(){
         String expected = "User already exists";
-        CreateUser createUser = new CreateUser("test@test.test", "Qwerty", "test");
+        CreateUser createUser = new CreateUser(email, password,name);
         ValidatableResponse createUserResponse = UserApi.postCreateUser(createUser);
-        createUserResponse.assertThat().statusCode(403).and().body("message", equalTo(expected));
+        accessToken = createUserResponse.assertThat().statusCode(200).extract().path("accessToken");
+        CreateUser createUserSimple = new CreateUser(email, password,name);
+        ValidatableResponse createUserResponseSimple = UserApi.postCreateUser(createUserSimple);
+        createUserResponseSimple.assertThat().statusCode(403).and().body("message", equalTo(expected));
     }
 
     @Test
